Refactor: add private folders to organize components part of SWM-101 related to #25
