+++ src/app/(nav)/question/searched-questions.tsx
@@ -1,4 +1,4 @@
-import { Question } from '@/app/types';
+import { Question } from '@/app/_lib/types';
 import QuestionCards from '@/components/question-card';
 
 const SearchedQuestions = ({ questions }: { questions: Question[] }) => {
