+++ src/app/(top)/workbookDetail/[workbookId]/_components/workbook-edit-buttons.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React, { useState } from 'react';
-import DeleteAlertDialog from '@/app/(top)/workbookDetail/[workbookId]/delete-alert-dialog';
+import DeleteAlertDialog from '@/app/(top)/workbookDetail/[workbookId]/_components/delete-alert-dialog';
 
 const WorkbookEditButtons = () => {
   const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState<boolean>(false);
