+++ src/app/(top)/workbookDetail/[workbookId]/_components/workbookDetail.stories.tsx
@@ -1,9 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { http, HttpResponse, delay } from 'msw';
-import { BASE_URL } from '@/app/constants';
-import { Dummy_Questions_In_Workbook, Dummy_Workbook } from '@/app/dummy';
+import { BASE_URL } from '@/app/_lib/constants';
+import { Dummy_Questions_In_Workbook, Dummy_Workbook } from '@/app/_lib/dummy';
 
-import Page from './page';
+import Page from '../page';
 
 const meta = {
   title: 'WorkbookDetail',
